requirements.txt
allure-pytest==2.9.45
pytest==6.2.5
selenium==4.1.0
pytest-rerunfailures==10.2

py==1.11.0
colorama~=0.4.4
pip==21.3.1
attrs==21.4.0
toml~=0.10.2
pluggy==1.0.0
iniconfig~=1.1.1
atomicwrites~=1.4.0
urllib3~=1.26.6
six~=1.16.0
pyparsing==3.0.6
setuptools==60.5.0
requests==2.27.1